+++ src/solver/RTree.hpp
@@ -15,7 +15,7 @@ class StatePool;
 
 class RTree : public StateIndex {
   public:
-    RTree(unsigned long nSDim, StatePool *statePool);
+    RTree(unsigned int nSDim, StatePool *statePool);
     virtual ~RTree() = default;
     RTree(RTree const &) = delete;
     RTree(RTree &&) = delete;
@@ -30,7 +30,7 @@ class RTree : public StateIndex {
             std::vector<double> highCorner);
 
   private:
-    unsigned long nSDim_;
+    unsigned int nSDim_;
     StatePool *statePool_;
     std::unique_ptr<Tools::PropertySet> properties_;
     std::unique_ptr<SpatialIndex::IStorageManager> storageManager_;
